+++ src/base/Processor.php
@@ -11,7 +11,9 @@
 namespace pennebaker\architect\base;
 
 use Craft;
-use craft\base\VolumeInterface;
+use craft\db\Query;
+use craft\db\Table;
+use craft\errors\SiteNotFoundException;
 use craft\models\FieldLayout;
 use craft\base\Field;
 
@@ -30,7 +32,8 @@ abstract class Processor implements ProcessorInterface
      *
      * @return FieldLayout
      */
-    public function createFieldLayout($item, $type) {
+    public function createFieldLayout($item, $type): FieldLayout
+    {
         $fieldLayout = new FieldLayout();
 
         if (isset($item['fieldLayout'])) {
@@ -67,7 +70,8 @@ abstract class Processor implements ProcessorInterface
      * @param array $obj
      * @return array
      */
-    public function stripNulls(array $obj) {
+    public function stripNulls(array $obj):array
+    {
         $allowedNulls = [
             'maxLevels'
         ];
@@ -82,16 +86,172 @@ abstract class Processor implements ProcessorInterface
     }
 
     /**
+     * @param $item
+     * @param string $type
+     * @param string $handle
+     * @param bool $prefix
+     */
+    public function mapService(&$item, $type, $handle, $prefix)
+    {
+        $service = null;
+        switch ($type) {
+            case 'asset':
+                $type = 'volume';
+                $service = Craft::$app->volumes->getVolumeByHandle($handle);
+                if (!$service) {
+                    $type = 'folder';
+                    $service = Craft::$app->assets->findFolder(['name' => $handle]);
+                }
+                break;
+            case 'volume':
+                $service = Craft::$app->volumes->getVolumeByHandle($handle);
+                break;
+            case 'folder':
+                $service = Craft::$app->assets->findFolder(['name' => $handle]);
+                break;
+            case 'section':
+                $service = Craft::$app->sections->getSectionByHandle($handle);
+                break;
+            case 'group':
+                $service = Craft::$app->categories->getGroupByHandle($handle);
+                if (!$service) {
+                    $service = Craft::$app->userGroups->getGroupByHandle($handle);
+                }
+                break;
+            case 'taggroup':
+                $service = Craft::$app->tags->getTagGroupByHandle($handle);
+                break;
+            case 'transform':
+                $service = Craft::$app->assetTransforms->getTransformByHandle($handle);
+                break;
+            default:
+                break;
+        }
+        if ($service) {
+            if ($prefix) {
+                $item = $type . ':' . $service->uid;
+            } else {
+                $item = $service->uid;
+            }
+        }
+    }
+
+    /**
+     * @param array|string|null $obj
+     * @param string $expectedType
+     * @param bool $prefix
+     */
+    public function map(&$obj, $expectedType, $prefix = true)
+    {
+        if (is_string($obj)) {
+            if ($obj !== '*' && $obj !== '') {
+                try {
+                    list($type, $handle) = explode(':', $obj);
+                } catch (\Exception $e) {
+                    $type = $expectedType;
+                    $handle = $obj;
+                }
+                $this->mapService($obj, $type, $handle, $prefix);
+            }
+        } else if (is_array($obj)) {
+            foreach ($obj as $k => &$item) {
+                try {
+                    list($type, $handle) = explode(':', $item);
+                } catch (\Exception $e) {
+                    $type = $expectedType;
+                    $handle = $item;
+                }
+                $this->mapService($item, $type, $handle, $prefix);
+            }
+        } else {
+            $obj = null;
+        }
+    }
+
+    /**
+     * @param $item
+     * @param string $type
+     * @param string $uid
+     */
+    public function unmapService(&$item, $type, $uid)
+    {
+        $service = null;
+        switch ($type) {
+            case 'volume':
+                $service = Craft::$app->volumes->getVolumeByUid($uid);
+                break;
+            case 'folder':
+                $service = Craft::$app->assets->getFolderByUid($uid);
+                break;
+            case 'section':
+                $service = Craft::$app->sections->getSectionByUid($uid);
+                break;
+            case 'group':
+                $service = Craft::$app->categories->getGroupByUid($uid);
+                if (!$service) {
+                    $service = Craft::$app->userGroups->getGroupByUid($uid);
+                }
+                break;
+            case 'taggroup':
+                $service = Craft::$app->tags->getTagGroupByUid($uid);
+                break;
+            case 'transform':
+                $service = Craft::$app->assetTransforms->getTransformByUid($uid);
+                break;
+            default:
+                break;
+        }
+        if ($service) {
+            try {
+                $handle = $service->handle;
+            } catch (\Exception $e) {
+                $handle = $service->name;
+            }
+            $item = $type . ':' . $handle;
+        }
+    }
+
+    /**
+     * @param array|string|null $obj
+     * @param \stdClass|null $expectedType
+     */
+    public function unmap(&$obj, $expectedType = null)
+    {
+        if (is_string($obj)) {
+            if ($obj !== '*' && $obj !== '') {
+                list($type, $uid) = explode(':', $obj);
+                $this->unmapService($obj, $type, $uid);
+            }
+        } else if (is_array($obj)) {
+            foreach ($obj as $k => &$item) {
+                try {
+                    list($type, $uid) = explode(':', $item);
+                } catch (\Exception $e) {
+                    $type = $expectedType;
+                    $uid = $item;
+                }
+                $this->unmapService($item, $type, $uid);
+            }
+        } else {
+            $obj = null;
+        }
+    }
+
+    /**
      * @param array|string $sites
      * @param string $prefix
      */
-    public function mapSites(&$sites, $prefix = '')
+    public function mapSites(&$sites, $prefix = '', $useIds = false)
     {
         if (is_array($sites)) {
             foreach ($sites as $k => $siteHandle) {
                 $site = Craft::$app->sites->getSiteByHandle($siteHandle);
                 if ($site) {
-                    $sites[$k] = $prefix . $site->id;
+                    if ($useIds) {
+                        $sites[$k] = $prefix . $site->id;
+                    } else {
+                        $sites[$k] = $prefix . $site->uid;
+                    }
                 } else {
                     unset($sites[$k]);
                 }
@@ -99,7 +259,11 @@ abstract class Processor implements ProcessorInterface
         } else {
             $site = Craft::$app->sites->getSiteByHandle($sites);
             if ($site) {
-                $sites = $prefix . $site->id;
+                if ($useIds) {
+                    $sites = $prefix . $site->id;
+                } else {
+                    $sites = $prefix . $site->uid;
+                }
             } else {
                 $sites = null;
             }
@@ -107,355 +271,38 @@ abstract class Processor implements ProcessorInterface
     }
 
     /**
-     * @param array|string $sites
-     * @param string $prefix
+     * @param array|string|null $sites
      */
-    public function unmapSites(&$sites, $prefix = '')
+    public function unmapSites(&$sites)
     {
         if (is_array($sites)) {
             foreach ($sites as $k => $siteRef) {
-                $siteId = substr($siteRef, strlen($prefix));
-                $site = Craft::$app->sites->getSiteById((int) $siteId);
-                if ($site) {
+                try {
+                    $site = Craft::$app->sites->getSiteByUid($siteRef);
                     $sites[$k] = $site->handle;
-                } else {
-                    unset($sites[$k]);
+                } catch (SiteNotFoundException $e) {
+                    $site = Craft::$app->sites->getSiteById($siteRef);
+                    if ($site) {
+                        $sites[$k] = $site->handle;
+                    } else {
+                        unset($sites[$k]);
+                    }
                 }
             }
         } else if (is_string($sites)) {
-            $siteId = substr($sites, strlen($prefix));
-            $site = Craft::$app->sites->getSiteById((int) $siteId);
-            if ($site) {
+            try {
+                $site = Craft::$app->sites->getSiteByUid($sites);
                 $sites = $site->handle;
-            } else {
-                $sites = null;
+            } catch (SiteNotFoundException $e) {
+                $site = Craft::$app->sites->getSiteById($sites);
+                if ($site) {
+                    $sites = $site->handle;
+                } else {
+                    $sites = null;
+                }
             }
         } else {
             $sites = null;
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function mapVolumeSources(&$sources, $prefix = '')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceHandle) {
-                $source = Craft::$app->volumes->getVolumeByHandle($sourceHandle);
-                if ($source) {
-                    $sources[$k] = $prefix . $source->id;
-                } else {
-                    unset($sources[$k]);
-                }
-            }
-        } else if (is_string($sources)) {
-            $source = Craft::$app->volumes->getVolumeByHandle($sources);
-            if ($source) {
-                $sources = $prefix . $source->id;
-            } else {
-                $sources = '*';
-            }
-        } else {
-            $sources = '*';
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function unmapVolumeSources(&$sources, $prefix = '')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                $sourceId = substr($sourceRef, strlen($prefix));
-                $source = Craft::$app->volumes->getVolumeById((int) $sourceId);
-                if ($source) {
-                    $sources[$k] = $source->handle;
-                } else {
-                    unset($sources[$k]);
-                }
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '') {
-            $sourceId = substr($sources, strlen($prefix));
-            $source = Craft::$app->volumes->getVolumeById((int) $sourceId);
-            if ($source) {
-                $sources = $source->handle;
-            } else {
-                $sources = '*';
-            }
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function mapFolderSources(&$sources, $prefix = 'folder:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceHandle) {
-                $volume = Craft::$app->volumes->getVolumeByHandle($sourceHandle);
-                $folder = Craft::$app->assets->getRootFolderByVolumeId($volume->id);
-                if ($folder) {
-                    $sources[$k] = $prefix . $folder->id;
-                } else {
-                    unset($sources[$k]);
-                }
-            }
-        } else if (is_string($sources)) {
-            /** @var VolumeInterface $source */
-            $volume = Craft::$app->volumes->getVolumeByHandle($sources);
-            if ($volume) {
-                $folder = Craft::$app->assets->getRootFolderByVolumeId($volume->id);
-            } else {
-                $folder = false;
-            }
-            if ($folder) {
-                $sources = $prefix . $folder->id;
-            } else {
-                $sources = '*';
-            }
-        } else {
-            $sources = '*';
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function unmapFolderSources(&$sources, $prefix = 'folder:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                $sourceId = substr($sourceRef, strlen($prefix));
-                $folder = Craft::$app->assets->getFolderById((int) $sourceId);
-                $volume = Craft::$app->volumes->getVolumeById((int) $folder->volumeId);
-                if ($volume) {
-                    $sources[$k] = $volume->handle;
-                } else {
-                    unset($sources[$k]);
-                }
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '') {
-            $sourceId = substr($sources, strlen($prefix));
-            $folder = Craft::$app->assets->getFolderById((int) $sourceId);
-            $volume = Craft::$app->volumes->getVolumeById((int) $folder->volumeId);
-            if ($volume) {
-                $sources = $volume->handle;
-            } else {
-                $sources = '*';
-            }
-        }
-    }
-
-    /**
-     * @param array|string $transforms
-     * @param string $prefix
-     */
-    public function mapAssetTransforms(&$transforms, $prefix = 'transform:')
-    {
-        if (is_array($transforms)) {
-            foreach ($transforms as $k => $transformHandle) {
-                $transform = Craft::$app->assetTransforms->getTransformByHandle($transformHandle);
-                if ($transform) {
-                    $transforms[$k] = $prefix . $transform->id;
-                } else {
-                    unset($transforms[$k]);
-                }
-            }
-        } else if (is_string($transforms)) {
-            $transform = Craft::$app->assetTransforms->getTransformByHandle($transforms);
-            if ($transform) {
-                $transforms = $prefix . $transform->id;
-            } else {
-                $transforms = '*';
-            }
-        } else {
-            $transforms = '*';
-        }
-    }
-
-    /**
-     * @param array|string $transforms
-     * @param string $prefix
-     */
-    public function unmapAssetTransforms(&$transforms, $prefix = 'transform:')
-    {
-        if (is_array($transforms)) {
-            foreach ($transforms as $k => $transformRef) {
-                $transformId = substr($transformRef, strlen($prefix));
-                $transform = Craft::$app->assetTransforms->getTransformById((int) $transformId);
-                if ($transform) {
-                    $transforms[$k] = $transform->handle;
-                } else {
-                    unset($transforms[$k]);
-                }
-            }
-        } else if (is_string($transforms)) {
-            $transformId = substr($transforms, strlen($prefix));
-            $transform = Craft::$app->assetTransforms->getTransformById((int) $transformId);
-            if ($transform) {
-                $transforms = $transform->handle;
-            } else {
-                $transforms = '*';
-            }
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function mapSectionSources(&$sources, $prefix = 'section:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceHandle) {
-                if ($sourceHandle !== 'singles') {
-                    $source = Craft::$app->sections->getSectionByHandle($sourceHandle);
-                    if ($source->type === 'single') {
-                        $sources[$k] = 'single:' . $source->id;
-                    } else {
-                        $sources[$k] = $prefix . $source->id;
-                    }
-                }
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '' && $sources !== 'singles') {
-            $source = Craft::$app->sections->getSectionByHandle((int) $sources);
-            $sources = $prefix . $source->id;
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function unmapSectionSources(&$sources, $prefix = 'section:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                if (substr($sourceRef, 0, strlen('single:')) === 'single:') {
-                    $sourceId = substr($sourceRef, strlen('single:'));
-                    $source = Craft::$app->sections->getSectionById((int) $sourceId);
-                    if ($source) {
-                        $sources[$k] = $source->handle;
-                    } else {
-                        unset($sources[$k]);
-                    }
-                } else if ($sourceRef !== 'singles') {
-                    $sourceId = substr($sourceRef, strlen($prefix));
-                    $source = Craft::$app->sections->getSectionById((int) $sourceId);
-                    if ($source) {
-                        $sources[$k] = $source->handle;
-                    } else {
-                        unset($sources[$k]);
-                    }
-                }
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '' && $sources !== 'singles') {
-            $source = Craft::$app->sections->getSectionById((int) $sources);
-            $sources = $source->handle;
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function mapCategorySources(&$sources, $prefix = 'group:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                $categoryGroup = Craft::$app->categories->getGroupByHandle($sourceRef);
-                $sources[$k] = $prefix . $categoryGroup->id;
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '') {
-            $categoryGroup = Craft::$app->categories->getGroupByHandle($sources);
-            $sources = $prefix . $categoryGroup->id;
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function unmapCategorySources(&$sources, $prefix = 'group:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                $sourceId = substr($sourceRef, strlen($prefix));
-                $categoryGroup = Craft::$app->categories->getGroupById((int) $sourceId);
-                $sources[$k] = $categoryGroup->handle;
-            }
-        } else if (isset($sources) && $sources !== '*' && $sources !== '') {
-            $sourceId = substr($sources, strlen($prefix));
-            $categoryGroup = Craft::$app->categories->getGroupById((int) $sourceId);
-            $sources = $categoryGroup->handle;
-        }
-    }
-
-    /**
-     * @param array|string $sourceHandle
-     * @param string $prefix
-     */
-    public function mapTagSource(&$sourceHandle, $prefix = 'taggroup:')
-    {
-        $source = Craft::$app->tags->getTagGroupByHandle($sourceHandle);
-        if ($source) {
-            $sourceHandle = $prefix . $source->id;
-        }
-    }
-
-    /**
-     * @param array|string $source
-     * @param string $prefix
-     */
-    public function unmapTagSource(&$source, $prefix = 'taggroup:')
-    {
-        $sourceId = substr($source, strlen($prefix));
-        $tagGroup = Craft::$app->tags->getTagGroupById((int) $sourceId);
-        if ($tagGroup) {
-            $source = $tagGroup->handle;
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function mapUserGroupSources(&$sources, $prefix = 'group:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceHandle) {
-                if ($sourceHandle !== 'admins') {
-                    $source = Craft::$app->userGroups->getGroupByHandle($sourceHandle);
-                    $sources[$k] = $prefix . $source->id;
-                }
-            }
-        } else {
-            $sources = '*';
-        }
-    }
-
-    /**
-     * @param array|string $sources
-     * @param string $prefix
-     */
-    public function unmapUserGroupSources(&$sources, $prefix = 'group:')
-    {
-        if (is_array($sources)) {
-            foreach ($sources as $k => $sourceRef) {
-                if ($sourceRef !== 'admins') {
-                    $sourceId = substr($sourceRef, strlen($prefix));
-                    $userGroup = Craft::$app->userGroups->getGroupById((int) $sourceId);
-                    $sources[$k] = $userGroup->handle;
-                }
-            }
-        } else {
-            $sources = '*';
         }
     }
 
@@ -509,7 +356,6 @@ abstract class Processor implements ProcessorInterface
      * @return array
      */
     public function exportFieldLayout($fieldLayout) {
-//        Craft::dump($fieldLayout);
         $fieldLayoutObj = [];
         $tabs = $fieldLayout->getTabs();
         usort($tabs, function($a, $b) {
@@ -530,7 +376,6 @@ abstract class Processor implements ProcessorInterface
      * @return array
      */
     public function exportRequiredFields($fieldLayout) {
-//        Craft::dump($fieldLayout);
         $requiredFieldsObj = [];
         $tabs = $fieldLayout->getTabs();
         usort($tabs, function($a, $b) {
@@ -582,22 +427,25 @@ abstract class Processor implements ProcessorInterface
                 try {
                     switch ($permissionMap[$permissionName]) {
                         case 'userGroup':
-                            $permissionHandle = Craft::$app->userGroups->getGroupById($permissionId)->handle;
+                            $permissionHandle = Craft::$app->userGroups->getGroupByUid($permissionId)->handle;
                             break;
                         case 'site':
-                            $permissionHandle = Craft::$app->sites->getSiteById($permissionId)->handle;
+                            $permissionHandle = Craft::$app->sites->getSiteByUid($permissionId)->handle;
                             break;
                         case 'section':
-                            $permissionHandle = Craft::$app->sections->getSectionById($permissionId)->handle;
+                            $permissionHandle = Craft::$app->sections->getSectionByUid($permissionId)->handle;
                             break;
                         case 'globalSet':
-                            $permissionHandle = Craft::$app->globals->getSetById($permissionId)->handle;
+                            $permissionHandle = (new Query())
+                                ->from([Table::GLOBALSETS])
+                                ->where(['uid' => $permissionId])
+                                ->one()['handle'];
                             break;
                         case 'category':
-                            $permissionHandle = Craft::$app->categories->getGroupById($permissionId)->handle;
+                            $permissionHandle = Craft::$app->categories->getGroupByUid($permissionId)->handle;
                             break;
                         case 'volume':
-                            $permissionHandle = Craft::$app->volumes->getVolumeById($permissionId)->handle;
+                            $permissionHandle = Craft::$app->volumes->getVolumeByUid($permissionId)->handle;
                             break;
                         default:
                             $permissionHandle = $permissionId;
@@ -620,22 +468,22 @@ abstract class Processor implements ProcessorInterface
                 try {
                     switch ($permissionMap[$permissionName]) {
                         case 'userGroup':
-                            $permissionId = Craft::$app->userGroups->getGroupByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->userGroups->getGroupByHandle($permissionHandle)->uid;
                             break;
                         case 'site':
-                            $permissionId = Craft::$app->sites->getSiteByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->sites->getSiteByHandle($permissionHandle)->uid;
                             break;
                         case 'section':
-                            $permissionId = Craft::$app->sections->getSectionByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->sections->getSectionByHandle($permissionHandle)->uid;
                             break;
                         case 'globalSet':
-                            $permissionId = Craft::$app->globals->getSetByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->globals->getSetByHandle($permissionHandle)->uid;
                             break;
                         case 'category':
-                            $permissionId = Craft::$app->categories->getGroupByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->categories->getGroupByHandle($permissionHandle)->uid;
                             break;
                         case 'volume':
-                            $permissionId = Craft::$app->volumes->getVolumeByHandle($permissionHandle)->id;
+                            $permissionId = Craft::$app->volumes->getVolumeByHandle($permissionHandle)->uid;
                             break;
                         default:
                             $permissionId = $permissionHandle;